ai-agents/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
pydantic==2.5.0
web3==6.11.3
redis==5.0.1
websockets==12.0
cryptography==41.0.7
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.88.1
httpx==0.25.2
python-dotenv==1.0.0

# AI and ML dependencies
google-generativeai==0.3.2
openai==1.3.7
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Notification services
sendgrid==6.10.0
twilio==8.10.3
firebase-admin==6.2.0

# ZK proof and crypto
py-ecc==6.0.0
eth-account==0.9.0
eth-keys==0.4.0